Ultimas correcciones by Imed

+++ src/politicasDeCancelacion/Intermedia.java
@@ -1,6 +1,7 @@
 package politicasDeCancelacion;
 
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 
 import observer.Inmueble;
 
@@ -15,7 +16,7 @@ public class Intermedia extends PoliticaDeCancelacion{
     	if (diasDeAntelacion >= 20) {
         		return 0; // Cancelación gratuita
     	} else if (diasDeAntelacion >= 10 && diasDeAntelacion < 20) {
-        		return reservaACancelar.getPrecioTotal() * 0.5; // Penalidad del 50%
+        		return (float) (reservaACancelar.getPrecioTotal() * 0.5); // Penalidad del 50%
     	} else {
         		return reservaACancelar.getPrecioTotal() ; // Penalidad total
     	}
